Move `FiniteBuffer` out to its own module
Just code motion.
